Upgraded tailwind and other npm deps.
