+++ docs/_includes/footer.html
@@ -1,6 +1,6 @@
 <section class="bg-black pt-10 pb-10">
     <div class="max-w-xs mx-auto text-center text-md leading-loose">
-        <img src="/static/frankdejonge.jpg" alt="Frank de Jonge" class="rounded-full h-20 w-20 border-grey-light border-2"/>
+        <img src="/static/frankdejonge.jpg" alt="Frank de Jonge" class="inline rounded-full h-20 w-20 border-grey-light border-2"/>
         <p class="text-grey">
             EventSauce is a project by <a class="text-white" target="_blank" href="https://frankdejonge.nl">Frank de Jonge</a>.
         </p>
