+++ docs/_layouts/default.html
@@ -24,7 +24,7 @@
     <!--<base href="//eventsauce.io/">-->
 </head>
 <body class="min-h-screen min-w-full pt-12 leading-normal bg-grey-lightest">
-<div class="h-16 fixed z-10 pin-t w-screen bg-white border-grey-light border-b">
+<div class="h-16 fixed z-10 top-0 w-screen bg-white border-grey-light border-b">
     <div class="max-w-lg lg:max-w-2xl mx-auto px-4 py-2">
         <div class="flex items-center flex-no-shrink pt-1">
             <a href="/" class="flex items-center">
@@ -36,7 +36,7 @@
             <div class="flex-grow"></div>
             <div class="flex-no-grow h-10 px-4 relative">
                 <input id="global-search" class="focus:outline-0 h-10 py-2 pl-8 pr-2 rounded border border-grey-lightest focus:border-grey-light bg-grey-lightest focus:bg-white focus:placeholder-white max-w-xs w-full appearance-none" type="text" placeholder="Search the docs" />
-                <span class="pointer-events-none absolute pin-y pin-l pl-6 flex items-center">
+                <span class="pointer-events-none absolute inset-y-0 left-0 pl-6 flex items-center">
                     <svg class="fill-current pointer-events-none text-grey w-4 h-4 border-transparent border" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M12.9 14.32a8 8 0 1 1 1.41-1.41l5.35 5.33-1.42 1.42-5.33-5.34zM8 14A6 6 0 1 0 8 2a6 6 0 0 0 0 12z"/></svg>
                 </span>
             </div>
