+++ docs/package.json
@@ -6,19 +6,20 @@
         "prod": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=webpack.config.js"
     },
     "devDependencies": {
-        "autoprefixer": "^8.6.5",
-        "color": "^3.1.0",
+        "autoprefixer": "^9.6.0",
+        "color": "^3.1.2",
         "cross-env": "^5.2.0",
-        "css-loader": "^0.28.11",
-        "extract-text-webpack-plugin": "^3.0.2",
+        "css-loader": "^3.0.0",
         "glob-all": "^3.1.0",
-        "optimize-css-assets-webpack-plugin": "^3.2.0",
-        "postcss": "^6.0.23",
-        "postcss-loader": "^2.1.6",
-        "purgecss-webpack-plugin": "^0.20.1",
-        "style-loader": "^0.19.0",
-        "tailwindcss": "^0.4.3",
-        "webpack": "^3.12.0"
+        "mini-css-extract-plugin": "^0.7.0",
+        "optimize-css-assets-webpack-plugin": "^5.0.1",
+        "postcss": "^7.0.17",
+        "postcss-loader": "^3.0.0",
+        "purgecss-webpack-plugin": "^1.5.0",
+        "style-loader": "^0.23.1",
+        "tailwindcss": "^1.0.4",
+        "webpack": "^4.34.0",
+        "webpack-cli": "^3.3.4"
     },
     "dependencies": {}
 }
