+++ docs/postcss.config.js
@@ -1,6 +1,6 @@
 module.exports = {
     plugins: [
         require('autoprefixer'),
-        require('tailwindcss')('./tailwind.js')
+        require('tailwindcss'),
     ]
 }
